fix stateManager
