+++ packages/client/tsconfig.json
@@ -1,6 +1,6 @@
 {
   "extends": ["@massmarket/tsconfig/base.json"],
-  "include": ["lib/**/*.ts", "test/*.ts"],
+  "include": ["src/**/*.ts", "test/*.ts"],
   "compilerOptions": {
     "outDir": "./dist"
   }
